app/requirements.txt
aws-cdk.core==1.34.0
aws-cdk.aws-s3==1.34.0